separated the Planner class from the UI (planning)
